delete unnecessary files to prepare for rebase
