+++ aiida_lsmo/workchains/isotherm_inflection.py
@@ -1,4 +1,4 @@
-"""IsothermInflection work chain"""
+"""A work chain."""
 
 import os
 import numpy as np
@@ -23,7 +23,7 @@ CifData = DataFactory('cif')  # pylint: disable=invalid-name
 ZeoppParameters = DataFactory('zeopp.parameters')  # pylint: disable=invalid-name
 
 # Deafault parameters
-ISOTHERMPARAMETERS_DEFAULT = {  #TODO: create IsothermParameters instead of Dict # pylint: disable=fixme
+PARAMETERS_DEFAULT = {
     "ff_framework": "UFF",  # (str) Forcefield of the structure.
     "ff_separate_interactions": False,  # (bool) Use "separate_interactions" in the FF builder.
     "ff_mixing_rule": "Lorentz-Berthelot",  # (string) Choose 'Lorentz-Berthelot' or 'Jorgensen'.
@@ -31,17 +31,17 @@ ISOTHERMPARAMETERS_DEFAULT = {  #TODO: create IsothermParameters instead of Dict
     "ff_shifted": False,  # (bool) Shift or truncate the potential at cutoff.
     "ff_cutoff": 12.0,  # (float) CutOff truncation for the VdW interactions (Angstrom).
     "temperature": 300,  # (float) Temperature of the simulation.
+    "zeopp_probe_scaling": 1.0,
     "zeopp_volpo_samples": int(1e5),  # (int) Number of samples for VOLPO calculation (per UC volume).
     "zeopp_block_samples": int(100),  # (int) Number of samples for BLOCK calculation (per A^3).
     "raspa_verbosity": 10,  # (int) Print stats every: number of cycles / raspa_verbosity.
     "raspa_widom_cycles": int(1e5),  # (int) Number of Widom cycles.
     "raspa_gcmc_init_cycles": int(1e3),  # (int) Number of GCMC initialization cycles.
     "raspa_gcmc_prod_cycles": int(1e4),  # (int) Number of GCMC production cycles.
-    "pressure_min": 0.01,  # (float) Min pressure in P/P0 TODO: MIN selected from the henry coefficient!
-    "pressure_max": 1.0,  # (float) Max pressure in P/P0  TODO: MAX always 1.0!
+    "pressure_min": 0.001,  # (float) Min pressure in P/P0 TODO: MIN selected from the henry coefficient!
+    "pressure_max": 1.0,  # (float) Max pressure in P/P0 
     "pressure_num": 20,  # (int) Number of pressure points considered, eqispaced in a log plot
     "pressure_list": None,  # (list) Pressure list in P/P0. If 'None' pressure points are computed from min/max/num.
-    "pressure_recompute": 3,  # (int) Number of pressure points to recompute from mid-density NVT, before the inflection
 }
 
 
@@ -71,7 +71,7 @@ def get_atomic_radii(isotparam):
 @calcfunction
 def get_zeopp_parameters(molecule_dict, isotparam):
     """Get the ZeoppParameters from the inputs of the workchain"""
-    probe_rad = molecule_dict["proberad"]
+    probe_rad = molecule_dict["proberad"] * isotparam['zeopp_probe_scaling']
     param_dict = {
         'ha': 'DEF',
         'volpo': [probe_rad, probe_rad, isotparam['zeopp_volpo_samples']],
@@ -121,7 +121,7 @@ def get_geometric_dict(zeopp_out, molecule):
 
 
 @calcfunction
-def get_output_parameters(geom_out, widom_out, **gcmc_dict):
+def get_output_parameters(pressures, geom_out, widom_out, **gcmc_dict):
     """Merge results from all the steps of the work chain.
     > geom_out (Dict) contains the output of Zeo++
     > widom_out (Dict) contains the output of Raspa's Widom insertions calculation
@@ -145,54 +145,54 @@ def get_output_parameters(geom_out, widom_out, **gcmc_dict):
         for label in widom_labels:
             out_dict.update({label: widom_out_mol[label]})
 
+        gcmc_out_mol = list(list(
+            gcmc_dict.values())[0]["framework_1"]["components"].values())[0]  #same for all, take just the first!
+
         isotherm = {
-            'pressure': [],
+            'pressure': pressures,
             'pressure_unit': 'bar',
-            'loading_absolute_average': [],
-            'loading_absolute_dev': [],
+            'loading_absolute_average_from_dil': [],
+            'loading_absolute_average_from_sat': [],
+            'loading_absolute_dev_from_dil': [],
+            'loading_absolute_dev_from_sat': [],
             'loading_absolute_unit': 'mol/kg',
-            'enthalpy_of_adsorption_average': [],
-            'enthalpy_of_adsorption_dev': [],
-            'enthalpy_of_adsorption_unit': 'kJ/mol'
+            'enthalpy_of_adsorption_average_from_dil': [],
+            'enthalpy_of_adsorption_average_from_sat': [],
+            'enthalpy_of_adsorption_dev_from_dil': [],
+            'enthalpy_of_adsorption_dev_from_sat': [],
+            'enthalpy_of_adsorption_unit': 'kJ/mol',
+            'conversion_factor_molec_uc_to_cm3stp_cm3': gcmc_out_mol['conversion_factor_molec_uc_to_cm3stp_cm3'],
+            'conversion_factor_molec_uc_to_mg_g': gcmc_out_mol['conversion_factor_molec_uc_to_mg_g'],
+            'conversion_factor_molec_uc_to_mol_kg': gcmc_out_mol['conversion_factor_molec_uc_to_mol_kg'],
         }
 
         conv_ener = 1.0 / 120.273  # K to kJ/mol
         conv_press = 1e-5  # Pa to bar
+        conv_load = isotherm["conversion_factor_molec_uc_to_mol_kg"]
 
-        all_out_keys = [k for k in gcmc_dict if k[:4] == "out_"]
+        for i in range(len(pressures)):
+            for inp, key_ext in [('dil', "_from_dil"), ('sat', "_from_sat")]:
+                gcmc_out = gcmc_dict['raspa_gcmc_{}_{}'.format(inp, i + 1)]["framework_1"]
+                gcmc_out_mol = list(gcmc_out["components"].values())[0]
 
-        for k in all_out_keys:
-            gcmc_inp = gcmc_dict["inp_" + k[4:]]
-            gcmc_out = gcmc_dict[k]["framework_1"]
-            gcmc_out_mol = list(gcmc_out["components"].values())[0]
-            conv_load = gcmc_out_mol["conversion_factor_molec_uc_to_mol_kg"]
+                for label in ['loading_absolute_average', 'loading_absolute_dev']:
+                    isotherm[label + key_ext].append(conv_load * gcmc_out_mol[label])
 
-            isotherm['pressure'].append(conv_press * gcmc_inp["System"]["framework_1"]["ExternalPressure"])
+                for label in ['enthalpy_of_adsorption_average', 'enthalpy_of_adsorption_dev']:
+                    if gcmc_out['general'][label]:
+                        isotherm[label + key_ext].append(conv_ener * gcmc_out['general'][label])
+                    else:  # when there are no particles and Raspa return Null enthalpy
+                        isotherm[label + key_ext].append(None)
 
-            for label in ['loading_absolute_average', 'loading_absolute_dev']:
-                isotherm[label].append(conv_load * gcmc_out_mol[label])
-
-            for label in ['enthalpy_of_adsorption_average', 'enthalpy_of_adsorption_dev']:
-                if gcmc_out['general'][label]:
-                    isotherm[label].append(conv_ener * gcmc_out['general'][label])
-                else:  # when there are no particles and Raspa return Null enthalpy
-                    isotherm[label].append(None)
-
-        out_dict.update({
-            "isotherm": isotherm,
-            'temperature': gcmc_inp["System"]["framework_1"]["ExternalTemperature"],
-            'temperature_unit': 'K',
-            'conversion_factor_molec_uc_to_cm3stp_cm3': gcmc_out_mol['conversion_factor_molec_uc_to_cm3stp_cm3'],
-            'conversion_factor_molec_uc_to_mg_g': gcmc_out_mol['conversion_factor_molec_uc_to_mg_g'],
-            'conversion_factor_molec_uc_to_mol_kg': gcmc_out_mol['conversion_factor_molec_uc_to_mol_kg'],
-        })
+        out_dict["isotherm"] = isotherm
 
     return Dict(dict=out_dict)
 
 
 class IsothermInflectionWorkChain(WorkChain):
-    """Workchain that computes volpo and blocking spheres: if accessible volpo>0
-    it also runs a raspa widom calculation for the Henry coefficient.
+    """A work chain to compute single component isotherms at adsorption and desorption:
+    GCMC calculations are run in parallell at all pressures, starting from the empty framework and the saturated system.
+    This workchain is useful to spot adsorption hysteresis.  
     """
 
     @classmethod
@@ -214,25 +214,12 @@ class IsothermInflectionWorkChain(WorkChain):
                    valid_type=Dict,
                    help='Parameters for the Isotherm workchain: will be merged with IsothermParameters_defaults.')
 
-        spec.input("geometric",
-                   valid_type=Dict,
-                   required=False,
-                   help='[Only used by IsothermMultiTempWorkChain] Already computed geometric properties')
-
         spec.outline(
             cls.setup,
             cls.run_zeopp,  # computes volpo and blocks
             if_(cls.should_run_widom)(  # if is porous
-                cls.run_raspa_widom,  # run raspa widom calculation
-                cls.init_raspa_gcmc,  # initializate setting for GCMC
-                while_(cls.should_run_another_gcmc)(  # new pressure
-                    cls.run_raspa_gcmc,  # run raspa GCMC calculation
-                    cls.check_inflection),
-                if_(cls.inflection_found)(
-                    cls.run_nvt_half_density,
-                    cls.run_gcmc_inflection_found,
-                    # cls.run_saturated_gcmc_all,
-                ),
+                cls.run_raspa_widom_and_sat,  # run raspa (1) widom calculation (2) NVT at qsat
+                cls.run_raspa_gcmc_from_dil_sat,  # run all calc from zero uptake and sat uptake
             ),
             cls.return_output_parameters,
         )
@@ -254,7 +241,7 @@ class IsothermInflectionWorkChain(WorkChain):
             self.ctx.molecule = self.inputs.molecule
 
         # Get the parameters Dict, merging defaults with user settings
-        self.ctx.parameters = aiida_dict_merge(Dict(dict=ISOTHERMPARAMETERS_DEFAULT), self.inputs.parameters)
+        self.ctx.parameters = aiida_dict_merge(Dict(dict=PARAMETERS_DEFAULT), self.inputs.parameters)
 
         # Get integer temperature in context for easy reports
         self.ctx.temperature = int(round(self.ctx.parameters['temperature']))
@@ -353,6 +340,7 @@ class IsothermInflectionWorkChain(WorkChain):
         # Check particular conditions and settings
         mult = check_resize_unit_cell(self.inputs.structure, 2 * self.ctx.parameters['ff_cutoff'])
         param["System"]["framework_1"]["UnitCells"] = "{} {} {}".format(mult[0], mult[1], mult[2])
+        self.ctx.number_uc = mult[0] * mult[1] * mult[2]
 
         if self.ctx.geom['Number_of_blocking_spheres'] > 0:
             param["Component"][self.ctx.molecule['name']]["BlockPocketsFileName"] = "block_file"
@@ -367,7 +355,7 @@ class IsothermInflectionWorkChain(WorkChain):
 
         return param
 
-    def _update_param_for_gcmc(self):
+    def _update_param_for_gcmc(self, number_of_molecules=0, swap_prob=0.5):
         """Update Raspa input parameter, from Widom to GCMC"""
 
         param = self.ctx.raspa_param
@@ -380,9 +368,9 @@ class IsothermInflectionWorkChain(WorkChain):
         param["Component"][self.ctx.molecule['name']].update({
             "WidomProbability": 0.0,
             "TranslationProbability": 1.0,
-            "ReinsertionProbability": 1.0,
-            "SwapProbability": 2.0,
-            "CreateNumberOfMolecules": 0,
+            "ReinsertionProbability": 0.5,
+            "SwapProbability": swap_prob,
+            "CreateNumberOfMolecules": number_of_molecules,
         })
         # Check particular conditions
         if not self.ctx.molecule['singlebead']:
@@ -393,27 +381,27 @@ class IsothermInflectionWorkChain(WorkChain):
 
         return param
 
-    def _update_param_for_nvt(self):
-        """Update Raspa input parameter, from GCMC to NVT"""
+    def _get_saturation_molecules(self):
+        """Compute the estimate of molecules at saturation by: pore_vol * lid_dens * number_uc."""
+        NAVO = 6.022E+23  # molec/mol
+        A3TOL = 1E-27  # L/A3
+        molar_dens = self.ctx.molecule['molsatdens']  #mol/l
+        pore_vol = self.ctx.zeopp.outputs.output_parameters["POAV_A^3"]  #A3/UC
+        molec_uc = molar_dens * NAVO * A3TOL * pore_vol  # molec/UC
+        molec_tot = molec_uc * self.ctx.number_uc
+        return int(molec_tot)
 
-        param = self.ctx.raspa_param
-        param["GeneralSettings"].update({
-            "NumberOfInitializationCycles": self.ctx.parameters['raspa_gcmc_init_cycles'],
-            "NumberOfCycles": self.ctx.parameters['raspa_gcmc_prod_cycles'],
-            "PrintPropertiesEvery": int(1e6),
-            "PrintEvery": self.ctx.parameters['raspa_gcmc_prod_cycles'] / self.ctx.parameters['raspa_verbosity']
-        })
+    def _get_mid_dens_molecules(self, raspa_calc_dil, raspa_calc_sat):
+        """Given a calculation at diluted and saturation condition, compute the total molecules at mid density."""
+        molec_uc_dil = list(raspa_calc_dil.outputs.output_parameters["framework_1"]
+                            ["components"].values())[0]["loading_absolute_average"]
+        molec_uc_sat = list(raspa_calc_sat.outputs.output_parameters["framework_1"]
+                            ["components"].values())[0]["loading_absolute_average"]
+        molec_uc_mid = max(0, (molec_uc_sat + molec_uc_dil) / 2)
+        molec_tot_mid = molec_uc_mid * self.ctx.number_uc
+        return int(molec_tot_mid)
 
-        last_uptake = self.ctx.raspa_gcmc[-1].outputs.output_parameters["framework_1"]["components"][
-            self.ctx.molecule['name']]['loading_absolute_average']
-
-        param["Component"][self.ctx.molecule['name']].update({
-            "SwapProbability": 0.0,
-            "CreateNumberOfMolecules": int(last_uptake * 0.5),  # halven!
-        })
-        return param
-
-    def run_raspa_widom(self):
+    def run_raspa_widom_and_sat(self):
         """Run a Widom calculation in Raspa."""
 
         # Initialize the input for raspa_base, which later will need only minor updates for GCMC
@@ -438,150 +426,54 @@ class IsothermInflectionWorkChain(WorkChain):
         self.report("Running Raspa Widom {} @ {}K for the Henry coefficient".format(self.ctx.molecule['name'],
                                                                                     self.ctx.temperature))
 
-        return ToContext(raspa_widom=running)
+        self.to_context(raspa_widom=running)
 
-    def init_raspa_gcmc(self):
-        """Choose the pressures we want to sample, report some details, and update settings for GCMC"""
+        # Run Raspa NVT at 90% Qsat
+        self.ctx.inp['metadata']['label'] = "RaspaNVT_sat"
+        self.ctx.inp['metadata']['call_link_label'] = "run_raspa_nvt_sat"
 
-        self.ctx.current_p_index = 0
+        nmolecules = int(0.90 * self._get_saturation_molecules())
+        self.ctx.raspa_param = self._update_param_for_gcmc(number_of_molecules=nmolecules, swap_prob=0.01)
+        self.ctx.inp['raspa']['parameters'] = Dict(dict=self.ctx.raspa_param)
+
+        running = self.submit(RaspaBaseWorkChain, **self.ctx.inp)
+
+        self.report("Running Raspa NVT with {} molecules".format(nmolecules))
+        self.to_context(raspa_nvt_sat=running)
+
+    def run_raspa_gcmc_from_dil_sat(self):
+        """For each pressure point, run GCMC calculation from both diluted and saturated initial conditions."""
+
+        self.report("Computed {} {}".format(
+            self.ctx.raspa_widom.outputs.output_parameters.pk,
+            self.ctx.raspa_nvt_sat.outputs.output_parameters.pk,
+        ))
 
         self.ctx.pressures = get_pressure_points(self.ctx.molecule, self.ctx.parameters)
-
-        self.report("{}: Kh(mol/kg/Pa)={:.2e} POAV(cm3/g)={:.3f} Qsat(mol/kg)={:.2f}".format(
-            self.ctx.molecule['name'],
-            list(self.ctx.raspa_widom.outputs['output_parameters']["framework_1"]["components"].values())[0]
-            ['henry_coefficient_average'], self.ctx.geom['POAV_cm^3/g'], self.ctx.geom['Estimated_saturation_loading']))
-        self.report("Now evaluating the isotherm {} @ {}K for {} pressure points".format(
-            self.ctx.molecule['name'], self.ctx.temperature, len(self.ctx.pressures)))
-
         self.ctx.raspa_param = self._update_param_for_gcmc()
 
-        self.ctx.loading_all = []
-        self.ctx.is_inflection_found = False
+        for inp in ['dil', 'sat']:
+            for i, pressure in enumerate(self.ctx.pressures):
+                # Update labels
+                self.ctx.inp['metadata']['label'] = "RaspaGCMC_{}_{}".format(inp, i + 1)
+                self.ctx.inp['metadata']['call_link_label'] = "run_raspa_gcmc_{}_{}".format(inp, i + 1)
 
-    def should_run_another_gcmc(self):
-        """Return False if all pressures have been computed or an inflection point was found."""
-        return self.ctx.current_p_index < len(self.ctx.pressures) and not self.ctx.is_inflection_found
+                self.ctx.raspa_param["System"]["framework_1"]['ExternalPressure'] = self.ctx.pressures[i] * 1e5
 
-    def inflection_found(self):
-        """To be used as variable for the logic. TODO: is there a better way?"""
-        return self.ctx.is_inflection_found
+                # Update parameters Dict
+                self.ctx.inp['raspa']['parameters'] = Dict(dict=self.ctx.raspa_param)
 
-    def check_inflection(self):
-        """Return True if an inflection point was found at the previous step."""
+                if inp == 'dil':
+                    pass
+                elif inp == 'sat':
+                    self.ctx.inp['raspa']['retrieved_parent_folder'] = self.ctx.raspa_nvt_sat.outputs.retrieved
 
-        # inflection point already found
-        if self.ctx.is_inflection_found:
-            return
+                # Create the calculation process, launch it and update pressure index
+                running = self.submit(RaspaBaseWorkChain, **self.ctx.inp)
+                self.report("Running Raspa GCMC {} {}".format(inp, i + 1))
 
-        if "conv_factor" not in self.ctx:
-            self.ctx.conv_factor = self.ctx.raspa_gcmc[-1].outputs.output_parameters["framework_1"]["components"][
-                self.ctx.molecule['name']]['conversion_factor_molec_uc_to_mol_kg']
-
-        loading_curr = self.ctx.conv_factor * self.ctx.raspa_gcmc[-1].outputs.output_parameters["framework_1"][
-            "components"][self.ctx.molecule['name']]['loading_absolute_average']
-
-        # check if inflection point present between the two last uptakes
-        if "loading_last" in self.ctx:
-            loading_diff = loading_curr - self.ctx.loading_last
-            if loading_diff > 0.4 * self.ctx.geom['Estimated_saturation_loading']:  # TODO: 40% to tune # pylint: disable=fixme
-                self.ctx.is_inflection_found = True
-                self.report("Inflection point found: {} > 0.4*{}".format(loading_diff,
-                                                                         self.ctx.geom['Estimated_saturation_loading']))
-
-        self.ctx.loading_last = loading_curr
-
-        return
-
-    def run_raspa_gcmc(self):
-        """Run a GCMC calculation in Raspa @ T,P. """
-
-        # Update labels
-        self.ctx.inp['metadata']['label'] = "RaspaGCMC_{}".format(self.ctx.current_p_index + 1)
-        self.ctx.inp['metadata']['call_link_label'] = "run_raspa_gcmc_{}".format(self.ctx.current_p_index + 1)
-
-        # Update pressure (NOTE: need to convert from bar to Pa)
-        self.ctx.raspa_param["System"]["framework_1"]['ExternalPressure'] = \
-            self.ctx.pressures[self.ctx.current_p_index] * 1e5
-
-        # Update parameters Dict
-        self.ctx.inp['raspa']['parameters'] = Dict(dict=self.ctx.raspa_param)
-
-        # Update restart (if present, i.e., if current_p_index>0)
-        if self.ctx.current_p_index > 0:
-            self.ctx.inp['raspa']['retrieved_parent_folder'] = self.ctx.raspa_gcmc[-1].outputs.retrieved
-
-        # Create the calculation process, launch it and update pressure index
-        running = self.submit(RaspaBaseWorkChain, **self.ctx.inp)
-        self.report("Running Raspa GCMC {} @ {}K/{:.3f}bar (pressure {} of {})".format(
-            self.ctx.molecule['name'], self.ctx.temperature, self.ctx.pressures[self.ctx.current_p_index],
-            self.ctx.current_p_index + 1, len(self.ctx.pressures)))
-        self.ctx.current_p_index += 1
-        return ToContext(raspa_gcmc=append_(running))
-
-    def run_nvt_half_density(self):
-        """Run a NVT calculation in Raspa, with half the loading of the previous GCMC."""
-
-        # Update labels
-        self.ctx.inp['metadata']['label'] = "RaspaNVT"
-        self.ctx.inp['metadata']['call_link_label'] = "run_raspa_nvt"
-
-        # Update parameters Dict
-        self.ctx.raspa_param = self._update_param_for_nvt()
-        self.ctx.inp['raspa']['parameters'] = Dict(dict=self.ctx.raspa_param)
-        del self.ctx.inp['raspa']['retrieved_parent_folder']
-
-        # Create the calculation process, launch it and update pressure index
-        running = self.submit(RaspaBaseWorkChain, **self.ctx.inp)
-        self.report("Running Raspa NVT @ half density")
-        return ToContext(raspa_nvt=running)
-
-    def run_gcmc_inflection_found(self):
-        """Run all the remaining GCMC calculations in parallel."""
-
-        self.ctx.raspa_param = self._update_param_for_gcmc()
-        self.ctx.raspa_param["GeneralSettings"]["NumberOfInitializationCycles"] = self.ctx.parameters[
-            'raspa_gcmc_prod_cycles']  # long initialization
-        self.ctx.inp['raspa']['retrieved_parent_folder'] = self.ctx.raspa_nvt.outputs.retrieved
-
-        # compute lower pressure points to check better the inflection point
-        inflection_press_index = self.ctx.current_p_index - 1
-        start_recomputing_index = inflection_press_index - self.ctx.parameters['pressure_recompute']
-        for i, pressure in enumerate(self.ctx.pressures[start_recomputing_index:inflection_press_index + 1]):
-
-            self.ctx.inp['metadata']['label'] = "RaspaGCMCNew_{}".format(i)
-            self.ctx.inp['metadata']['call_link_label'] = "run_raspa_gcmc_new_{}".format(i)
-
-            self.ctx.raspa_param["System"]["framework_1"]['ExternalPressure'] = pressure * 1e5
-            self.ctx.inp['raspa']['parameters'] = Dict(dict=self.ctx.raspa_param)
-
-            # Create the calculation process, launch it and update pressure index
-            running = self.submit(RaspaBaseWorkChain, **self.ctx.inp)
-            self.report("Submit Raspa GCMC NEW @ {} bar".format(round(pressure, 2)))
-            self.to_context(raspa_gcmc=append_(running))
-
-        # compute the remaining higher pressures up to p=p0
-        # NOTE.1: these are expensive calculation with saturated gas. I don't want to be too precise. Consider CFCMC.
-        # NOTE.2: if there are two inflection points (pores of different size) this workchain does not work well.
-
-        self.ctx.inp['raspa']['retrieved_parent_folder'] = self.ctx.raspa_gcmc[inflection_press_index].outputs.retrieved
-
-        for i in [1, 2]:
-            if i == 1:  # mid point between last GCMC and p0
-                pressure = 0.5 * (self.ctx.pressures[inflection_press_index] + self.ctx.molecule['pressure_zero'])
-            elif i == 2:  # p=p0
-                pressure = self.ctx.molecule['pressure_zero']
-
-            self.ctx.inp['metadata']['label'] = "RaspaGCMCSat_{}".format(i)
-            self.ctx.inp['metadata']['call_link_label'] = "run_raspa_gcmc_sat_{}".format(i)
-
-            self.ctx.raspa_param["System"]["framework_1"]['ExternalPressure'] = pressure * 1e5
-            self.ctx.inp['raspa']['parameters'] = Dict(dict=self.ctx.raspa_param)
-
-            # Create the calculation process, launch it and update pressure index
-            running = self.submit(RaspaBaseWorkChain, **self.ctx.inp)
-            self.report("Submit Raspa GCMC Sat @ {} bar".format(round(pressure, 2)))
-            self.to_context(raspa_gcmc=append_(running))
+                key_ctx = 'raspa_gcmc_{}_{}'.format(inp, i + 1)
+                self.to_context(**{key_ctx: running})
 
     def return_output_parameters(self):
         """Merge all the parameters into output_parameters, depending on is_porous and is_kh_ehough."""
@@ -589,13 +481,18 @@ class IsothermInflectionWorkChain(WorkChain):
         gcmc_dict = {}
         if self.ctx.geom['is_porous'] and not self.ctx.multitemp_mode == 'run_geom_only':
             widom_out = self.ctx.raspa_widom.outputs.output_parameters
-            for calc in self.ctx.raspa_gcmc:
-                gcmc_dict["inp_" + calc.label] = calc.inputs.raspa__parameters
-                gcmc_dict["out_" + calc.label] = calc.outputs.output_parameters
+            for key in self.ctx.keys():
+                if 'raspa_gcmc' in key:
+                    gcmc_dict[key] = self.ctx[key].outputs.output_parameters
         else:
             widom_out = None
 
-        self.out("output_parameters", get_output_parameters(geom_out=self.ctx.geom, widom_out=widom_out, **gcmc_dict))
+        self.out(
+            "output_parameters",
+            get_output_parameters(pressures=self.ctx.pressures,
+                                  geom_out=self.ctx.geom,
+                                  widom_out=widom_out,
+                                  **gcmc_dict))
 
         if not self.ctx.multitemp_mode == 'run_geom_only':
             self.report("IsothermInflection {} @ {}K computed: ouput Dict<{}>".format(
